examples/hello: review app organization
